perf: optimize trait implementations and reorganize module structure
- Move traits.rs from src/ to src/app_path/ for better code organization
- Optimize all trait implementations to use direct field access instead of method calls
- Remove 6 redundant methods (exists, file_name, file_stem, extension, is_dir, is_file)
  - All functionality remains available via Deref<Target=Path>
- Add into_inner() method for zero-cost PathBuf extraction
- Optimize From<AppPath> trait implementations to move instead of clone
- Eliminate unnecessary function call overhead in hot paths

Performance improvements:
- Direct field access in Display, AsRef, PartialEq, Ord, Hash, Deref, Borrow traits
- Zero-cost conversions in From<AppPath> for PathBuf and OsString
- Cleaner API surface while preserving all functionality

All 141 unit tests + 60 doc tests pass - no breaking changes.
